chore: rename `UpdatesConfig` => `ClientUpdatesConfig`
