+++ src/models/public-api.ts
@@ -2,4 +2,4 @@ export * from './fast-alfred-config.model'
 export * from './alfred-list-item.model'
 export * from './alfred-script-filter.model'
 export * from './cache-item.model'
-export { UpdatesConfig, UpdatesFetcher, UpdatesFetcherResponse } from './updates-config.model'
+export { ClientUpdatesConfig, UpdatesFetcher, UpdatesFetcherResponse } from './client-updates-config.model'
